[fix] fix import in utils follwoing refactoring & allow to path one folder when loading a maddpg agent
